requirements.txt
Flask==1.0.2
python-dotenv
flickrapi
requests
newsapi-python
tweepy
instagram-scraper